requirements.txt
# Core Dependencies
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Unsloth für effizientes Training
# unsloth @ git+https://github.com/unslothai/unsloth.git

# Hugging Face Ecosystem
transformers>=4.36.0
accelerate>=0.25.0
peft>=0.7.0
bitsandbytes>=0.41.0
datasets>=2.15.0

# Training & RL
trl>=0.7.10
tensorboard>=2.15.0
wandb>=0.16.0

# LangChain & LangGraph
langgraph>=0.0.30
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10

# CrewAI (optional für alternative Orchestrierung)
crewai>=0.1.0

# LLM Providers
anthropic>=0.8.0
openai>=1.6.0

# Verdict - LLM as Judge
# verdict @ git+https://github.com/haizelabs/verdict.git

# Data & Finance
yfinance>=0.2.33
pandas>=2.1.0
numpy>=1.24.0
pandas-ta>=0.3.14b

# API Clients
requests>=2.31.0
aiohttp>=3.9.0
finnhub-python>=2.4.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pyyaml>=6.0

# Database
sqlalchemy>=2.0.0
sqlite3worker>=1.1.7

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code Quality
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0

# Monitoring
prometheus-client>=0.19.0

# Utilities
tqdm>=4.66.0
rich>=13.7.0
loguru>=0.7.0

# Serialization
orjson>=3.9.0
msgpack>=1.0.0

# Math & Stats
scipy>=1.11.0
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Contextual Bandits
scikit-multiflow>=0.5.3

# NLP
spacy>=3.7.0
nltk>=3.8.0

# API Framework (für Production-Deployment)
fastapi>=0.108.0
uvicorn[standard]>=0.25.0

# Caching
redis>=5.0.0
diskcache>=5.6.0
